+++ R/utils-yaml.R
@@ -187,7 +187,11 @@ create_pkgdown_yaml <- function(path) {
       NULL     
     )
   )
-  structure(yaml::yaml.load(yaml, eval.expr = FALSE), header = get_information_header(yaml))
+  rendered <- yaml::yaml.load(yaml, eval.expr = FALSE)
+  items <- names(rendered$template$params)
+  rendered$template$params <- c(rendered$template$params, 
+    c(usr[!names(usr) %in% items]))
+  structure(rendered, header = get_information_header(yaml))
 }
 
 update_site_timestamp <- function(path) {
@@ -205,14 +209,31 @@ get_navbar_info <- function(i) {
   )
 }
 
+known_yaml_keys <- c(
+  "title",
+  "carpentry",
+  "life_cycle",
+  "license",
+  "source",
+  "branch",
+  "contact",
+  "created",
+  "keywords"
+)
+
+known_yaml_items <- c(
+  known_yaml_keys,
+  "episodes",
+  "instructors",
+  "learners",
+  "profiles"
+)
+
+
 quote_config_items <- function(yaml) {
-  yaml$title      <- siQuote(yaml$title)
-  yaml$carpentry  <- siQuote(yaml$carpentry)
-  yaml$life_cycle <- siQuote(yaml$life_cycle)
-  yaml$license    <- siQuote(yaml$license)
-  yaml$source     <- siQuote(yaml$source)
-  yaml$branch     <- siQuote(yaml$branch)
-  yaml$contact    <- siQuote(yaml$contact)
+  for (i in known_yaml_keys) {
+    yaml[[i]] <- siQuote(yaml[[i]])
+  }
   yaml
 }