+++ tests/testthat/test-set_dropdown.R
@@ -29,6 +29,18 @@ cli::test_that_cli("set_config() will write items", {
     transform = function(s) mask_tmpdir(s, dirname(tmp)))
 })
 
+
+
+test_that("custom keys can be modified by set_config()", {
+  writeLines(c("test-key: 'hey!'", readLines(fs::path(tmp, "config.yaml"))),
+    fs::path(tmp, "config.yaml"))
+  expect_equal(get_config(tmp)[["test-key"]], "hey!")
+  expect_snapshot(set_config(c("test-key" = "!yeh"), path = tmp, write = TRUE),
+    transform = function(s) mask_tmpdir(s, dirname(tmp)))
+  expect_equal(get_config(tmp)[["test-key"]], "!yeh")
+  expect_equal(readLines(fs::path(tmp, "config.yaml"), n = 1L), "test-key: '!yeh'")
+})
+
 test_that("schedule is empty by default", {
 
   cfg <- get_config(tmp)
@@ -41,6 +53,10 @@ test_that("schedule is empty by default", {
   # the config files should be unchanged from the schedule
   no_episodes <- names(cfg)[names(cfg) != "episodes"]
   expect_equal(cfg[no_episodes], get_config(tmp)[no_episodes])
+
+  # the config file now has the `test-key` at the bottom
+  cfgtxt <- readLines(fs::path(tmp, "config.yaml"))
+  expect_equal(cfgtxt[length(cfgtxt) - 1L], "test-key: '!yeh'")
 
 })